cros_async: Delete RunnableQueue::set_waker
Previously each executor had to call RunnableQueue::set_waker at the
beginning of every loop before it started executing futures, which was a
bit tedious.  This call was needed so that the RunnableQueue would wake
up the executor if a new future was scheduled.

Instead move responsibility for waking the executor when a future is
scheduled back to the executor.  This lets us get rid of the set_waker
method and is arguably a better separation of responsibilities.

BUG=none
TEST=unit tests

Change-Id: Ica46437f78c822d87096eaa215954d118c6578f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosvm/+/2760377
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Reviewed-by: Noah Gold <nkgold@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>